Fix #889: fixed crash due to using ::getUnit without checking return value for NULL
::getUnit has NULL as a valid return value, yet most of the code will use the returned unit pointer
without testing if it's NULL. This has resulted in crashes in the past and it did it again.

This commit has reviewed all 282 calls to ::getUnit and protected each from crashing due to NULL pointers.

Note: it can be argued that it's a bug if we get a NULL pointer, but we do and the end result are crashes.

Two calls have yet to be protected against NULL pointers. Both are in pathfinding.

+++ Project Files/DLLSources/CvPlayer.cpp
@@ -9789,7 +9789,7 @@ void CvPlayer::updateGroupCycle(CvUnit* pUnit)
 		pLoopUnit = ::getUnit(pUnitNode->m_data);
 		pUnitNode = pPlot->nextUnitNode(pUnitNode);
 
-		if (pLoopUnit->isGroupHead())
+		if (pLoopUnit != NULL && pLoopUnit->isGroupHead())
 		{
 			if (pLoopUnit != pUnit)
 			{
@@ -11786,7 +11786,7 @@ void CvPlayer::doAdvancedStartAction(AdvancedStartActionTypes eAction, Coordinat
 						CvUnit* pLoopUnit = ::getUnit(pUnitNode->m_data);
 						pUnitNode = pPlot->nextUnitNode(pUnitNode);
 
-						if (pLoopUnit->getUnitType() == eUnit)
+						if (pLoopUnit != NULL && pLoopUnit->getUnitType() == eUnit)
 						{
 							pLoopUnit->kill(false);
 							changeAdvancedStartPoints(iCost);
@@ -11801,10 +11801,13 @@ void CvPlayer::doAdvancedStartAction(AdvancedStartActionTypes eAction, Coordinat
 				{
 					CvUnit* pUnit = ::getUnit(pUnitNode->m_data);
 
-					iCost = getAdvancedStartUnitCost(pUnit->getUnitType(), false);
-					FAssertMsg(iCost != -1, "If this is -1 then that means it's going to try to delete a unit which shouldn't exist");
-					pUnit->kill(false);
-					changeAdvancedStartPoints(iCost);
+					if (pUnit != NULL)
+					{
+						iCost = getAdvancedStartUnitCost(pUnit->getUnitType(), false);
+						FAssertMsg(iCost != -1, "If this is -1 then that means it's going to try to delete a unit which shouldn't exist");
+						pUnit->kill(false);
+						changeAdvancedStartPoints(iCost);
+					}
 				}
 			}
 
@@ -12303,7 +12306,7 @@ int CvPlayer::getAdvancedStartUnitCost(UnitTypes eUnit, bool bAdd, CvPlot* pPlot
 				CvUnit* pLoopUnit = ::getUnit(pUnitNode->m_data);
 				pUnitNode = pPlot->nextUnitNode(pUnitNode);
 
-				if (pLoopUnit->getUnitType() == eUnit)
+				if (pLoopUnit != NULL && pLoopUnit->getUnitType() == eUnit)
 				{
 					bUnitFound = true;
 				}
